+++ packages/medusa/src/policies/order.ts
@@ -1,9 +1,5 @@
-import {
-  definePolicies,
-  PolicyDefinition,
-  toPascalCase,
-} from "@medusajs/framework/utils"
-import { defaultPolicyOperations } from "../utils/default-policy-operations"
+import { definePolicies } from "@medusajs/framework/utils"
+import { generateResourcePolicies } from "../utils"
 
 const orderResources = [
   "order",
@@ -16,19 +12,6 @@ const orderResources = [
   "return_reason",
 ]
 
-const policies: PolicyDefinition[] = []
-
-for (const resource of orderResources) {
-  for (const operation of defaultPolicyOperations) {
-    const policyName = toPascalCase(operation) + toPascalCase(resource)
-
-    policies.push({
-      name: policyName,
-      resource: resource,
-      operation: operation,
-      description: `${toPascalCase(operation)} ${resource.replace(/_/g, " ")}`,
-    })
-  }
-}
-
-export const orderPolicies = definePolicies(policies)
+export const orderPolicies = definePolicies(
+  generateResourcePolicies(orderResources)
+)