+++ packages/medusa/src/policies/payment.ts
@@ -1,9 +1,5 @@
-import {
-  definePolicies,
-  PolicyDefinition,
-  toPascalCase,
-} from "@medusajs/framework/utils"
-import { defaultPolicyOperations } from "../utils/default-policy-operations"
+import { definePolicies } from "@medusajs/framework/utils"
+import { generateResourcePolicies } from "../utils"
 
 const paymentResources = [
   "payment",
@@ -13,19 +9,6 @@ const paymentResources = [
   "refund_reason",
 ]
 
-const policies: PolicyDefinition[] = []
-
-for (const resource of paymentResources) {
-  for (const operation of defaultPolicyOperations) {
-    const policyName = toPascalCase(operation) + toPascalCase(resource)
-
-    policies.push({
-      name: policyName,
-      resource: resource,
-      operation: operation,
-      description: `${toPascalCase(operation)} ${resource.replace(/_/g, " ")}`,
-    })
-  }
-}
-
-export const paymentPolicies = definePolicies(policies)
+export const paymentPolicies = definePolicies(
+  generateResourcePolicies(paymentResources)
+)