+++ packages/modules/auth/src/models/provider-identity.ts
@@ -1,4 +1,6 @@
 import { model } from "@medusajs/framework/utils"
+import { AuthPasswordResetToken } from "./auth-password-reset-token"
+import { AuthVerificationToken } from "./auth-verification-token"
 import { AuthIdentity } from "./auth-identity"
 
 export const ProviderIdentity = model
@@ -8,6 +10,12 @@ export const ProviderIdentity = model
     provider: model.text(),
     auth_identity: model.belongsTo(() => AuthIdentity, {
       mappedBy: "provider_identities",
+    }),
+    verification_tokens: model.hasMany(() => AuthVerificationToken, {
+      mappedBy: "provider_identity",
+    }),
+    password_reset_tokens: model.hasMany(() => AuthPasswordResetToken, {
+      mappedBy: "provider_identity",
     }),
     user_metadata: model.json().nullable(),
     provider_metadata: model.json().nullable(),
@@ -19,3 +27,6 @@ export const ProviderIdentity = model
       unique: true,
     },
   ])
+  .cascades({
+    delete: ["verification_tokens", "password_reset_tokens"],
+  })